requirements.txt
flask>=3.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
requests>=2.31.0
werkzeug>=3.0.0
